# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: omu-rest
  description: 'Laravel製RESTful APIサーバー。著者: [connect0459](https://github.com/connect0459)'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api'
    description: ローカルホストのAPIエンドポイント
  -
    url: 'https://example.studyfortwo.org/omu-rest/public/api'
    description: lolipopのAPIエンドポイント
paths:
...
  /news:
    get:
      tags:
        - news
      summary: ニュースのリストを取得する
      operationId: 21b84dbdd75fd20b9c9ddb9b9e68f0a1
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/News'
    post:
      tags:
        - news
      summary: 新しいニュースを作成する
      operationId: d6a12c2e3b4ea9ab588ba9b1cb14748f
      requestBody:
        description: ニュースのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/News'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/News'
  '/news/{id}':
    get:
      tags:
        - news
      summary: 特定のIDでニュースを取得する
      operationId: eb815321a4dd395639d4aeef25d9a45c
      parameters:
        -
          name: id
          in: path
          description: ニュースのID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/News'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - news
      summary: 特定のIDでニュースを更新する
      operationId: 25ea273290c6cb379ab85e08ff4fadfc
      parameters:
        -
          name: id
          in: path
          description: ニュースのID
          required: true
          schema:
            type: string
      requestBody:
        description: ニュースのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/News'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/News'
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - news
      summary: 特定のIDでニュースを削除する
      operationId: bbe79a918bfbb58bc46ed74b43e9b986
      parameters:
        -
          name: id
          in: path
          description: ニュースのID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
  /orders_customers:
    get:
      tags:
        - orders_customers
      summary: 注文顧客のリストを取得する
      operationId: 5217a452519ec6ef916c6a3e2f4070dc
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderCustomer'
    post:
      tags:
        - orders_customers
      summary: 新しい注文顧客を作成する
      operationId: 75e0485900f170d2424fdeadf59b0752
      requestBody:
        description: OrderCustomerのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCustomer'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/OrderCustomer'
  '/orders_customers/{id}':
    get:
      tags:
        - orders_customers
      summary: 特定のIDで注文顧客を取得する
      operationId: 4dd012aac3be308644f1e2e152f92c19
      parameters:
        -
          name: id
          in: path
          description: 注文顧客のID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderCustomer'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - orders_customers
      summary: 特定のIDで注文顧客を更新する
      operationId: a0a94057f0311e2193b5897f80221470
      parameters:
        -
          name: id
          in: path
          description: 注文顧客のID
          required: true
          schema:
            type: string
      requestBody:
        description: OrderCustomerのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCustomer'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/OrderCustomer'
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - orders_customers
      summary: 特定のIDで注文顧客を削除する
      operationId: 5608eff2ef8984e0b7447765f27595aa
      parameters:
        -
          name: id
          in: path
          description: 注文顧客のID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
  /orders_items:
    get:
      tags:
        - orders_items
      summary: 注文商品のリストを取得する
      operationId: 89fde6ad08076b992c521fb7c680e29e
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderItem'
    post:
      tags:
        - orders_items
      summary: 新しい注文商品を作成する
      operationId: b2909cc3849277b1d5345ccb4205c71d
      requestBody:
        description: OrderItemのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItem'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/OrderItem'
  '/orders_items/{id}':
    get:
      tags:
        - orders_items
      summary: 特定のIDで注文商品を取得する
      operationId: 0b2a7a56e7d5c2c49b7d11bdb0f48f5c
      parameters:
        -
          name: id
          in: path
          description: 注文商品のID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderItem'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - orders_items
      summary: 特定のIDで注文商品を更新する
      operationId: 8bee2e95aba24db3af9b2bc7b5a53f0d
      parameters:
        -
          name: id
          in: path
          description: 注文商品のID
          required: true
          schema:
            type: string
      requestBody:
        description: OrderItemのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItem'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/OrderItem'
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - orders_items
      summary: 特定のIDで注文商品を削除する
      operationId: 6d397dbc3afd3a3f44e4cc08820e0921
      parameters:
        -
          name: id
          in: path
          description: 注文商品のID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
...
components:
  schemas:
...
    News:
      description: 'News Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        title:
          description: ニュースのタイトル
          type: string
          format: string
          example: サイトの稼働が開始しました。
        detail:
          description: ニュースの記事
          type: string
          format: string
          example: 2023年6月20日からサイトを稼働しております。
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
    OrderCustomer:
      description: 'OrderCustomer Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        type_branch_id:
          description: 支部番号（types_branchesテーブルからの外部キー制約）
          type: integer
          format: int64
          example: '1'
        numbering:
          description: 整理番号
          type: string
          format: string
          example: 231B001
        name1:
          description: 名前
          type: string
          format: string
          example: 中岡暉
        name2:
          description: 'フリガナ or ローマ字'
          type: string
          format: string
          example: ナカオカアキラ
        university:
          description: 大学
          type: string
          format: string
          example: ○○大学
        grade:
          description: 学年
          type: string
          format: string
          example: 1年
        belong:
          description: 学科など
          type: string
          format: string
          example: 工学部
        email:
          description: メールアドレス
          type: string
          format: string
          example: example@gmail.com
        total:
          description: 請求合計金額
          type: integer
          format: int64
          example: '1000'
        type_receive_id:
          description: 受取り方法分類id（types_receivesテーブルからの外部キー制約）
          type: integer
          format: int64
          example: '1'
        receive_date:
          description: 受取り日時
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        type_payment_id:
          description: 支払い方法分類id（types_paymentsテーブルからの外部キー制約）
          type: integer
          format: int64
          example: '1'
        payment_due_date:
          description: 支払い期限
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        notes:
          description: 備考・質問
          type: string
          format: string
          example: 受け取り日時は後から変更できますか？
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
    OrderItem:
      description: 'OrderItem Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        order_customer_id:
          description: 外部キー制約（orders_customersテーブル）
          type: integer
          format: int64
          example: '1'
        numbering:
          description: 整理番号
          type: string
          format: string
          example: 231B001
        type_order_state_id:
          description: 予約の処理状態を表すステータス。types_orders_statesテーブルから取得
          type: integer
          format: int64
          example: '1'
        book_info_id:
          description: 外部キー制約（books_info）
          type: integer
          format: int64
          example: '1'
        isbn:
          description: 'ISBN-13 or ISBN-10'
          type: string
          format: string
          example: '9784908434266'
        title:
          description: タイトル
          type: string
          format: string
          example: これだけは知っておこう!情報リテラシー
        sale_price:
          description: 販売額
          type: integer
          format: int64
          example: '500'
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: api_token
      in: header